Move JSONSkinLoader and JsonSkin definition classes
