+++ src/bms/player/beatoraja/skin/lua/LuaSkinLoader.java
@@ -4,6 +4,8 @@ import bms.player.beatoraja.Config;
 import bms.player.beatoraja.MainState;
 import bms.player.beatoraja.SkinConfig;
 import bms.player.beatoraja.skin.*;
+import bms.player.beatoraja.skin.json.JSONSkinLoader;
+import bms.player.beatoraja.skin.json.JsonSkin;
 import bms.player.beatoraja.skin.property.*;
 import com.badlogic.gdx.utils.reflect.ClassReflection;
 import com.badlogic.gdx.utils.reflect.Field;
@@ -34,7 +36,7 @@ public class LuaSkinLoader extends JSONSkinLoader {
 		try {
 			lua.setDirectory(p.getParent());
 			LuaValue value = lua.execFile(p);
-			sk = fromLuaValue(JsonSkin.class, value);
+			sk = fromLuaValue(JsonSkin.Skin.class, value);
 			header = loadJsonSkinHeader(sk, p);
 		} catch (Throwable e) {
 			e.printStackTrace();
@@ -57,7 +59,7 @@ public class LuaSkinLoader extends JSONSkinLoader {
 				return getPath(p.getParent().toString() + "/" + path, filemap).getPath();
 			});
 			LuaValue value = lua.execFile(p);
-			sk = fromLuaValue(JsonSkin.class, value);
+			sk = fromLuaValue(JsonSkin.Skin.class, value);
 			skin = loadJsonSkin(header, sk, type, property, p);
 		} catch (Throwable e) {
 			e.printStackTrace();
